[flink] Extract RefreshBlacklist from FileStoreLookupFunction
